model start field correction, service + error handling and test updates

+++ cashierApp/src/main/java/alimentation/cashierApp/services/EmployeeService.java
@@ -3,9 +3,11 @@ package alimentation.cashierApp.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import alimentation.cashierApp.dao.EmployeeRepository;
+import alimentation.cashierApp.exception.CashierAppException;
 import alimentation.cashierApp.models.Employee;
 
 @Service
@@ -16,27 +18,31 @@ public class EmployeeService {
 
 
     // Add methods for employee-related operations here
-    Iterable<Employee> getAllEmployees(){
+    public Iterable<Employee> getAllEmployees(){
         return employeeRepository.findAll();
     }
 
-    Employee getEmployeeById(int id){
+    public Employee getEmployeeById(int id){
         Optional<Employee> response = employeeRepository.findById(id);
-        if (response.isPresent()) {
-            return response.get();
+        if (response.isEmpty()) {
+            throw new CashierAppException(HttpStatus.NOT_FOUND, "Employee not found");
         }
-        return null;
+        return response.get();
     }
 
-    void addEmployee(Employee employee){
-        employeeRepository.save(employee);
+    public Employee createEmployee(Employee employee){
+        return employeeRepository.save(employee);
     }
 
-    void updateEmployee(Employee employee){
-        employeeRepository.save(employee);
+    public Employee updateEmployee(Employee employee){
+        return employeeRepository.save(employee);
     }
 
-    void deleteEmployee(int id){
+    public void deleteEmployee(int id){
+        Optional<Employee> target = employeeRepository.findById(id);
+        if (target.isEmpty()) {
+            throw new CashierAppException(HttpStatus.NOT_FOUND, "Employee not found");
+        }
         employeeRepository.deleteById(id);
     }
 }
