+++ cashierApp/src/main/java/alimentation/cashierApp/services/ReportService.java
@@ -3,9 +3,11 @@ package alimentation.cashierApp.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import alimentation.cashierApp.dao.ReportRepository;
+import alimentation.cashierApp.exception.CashierAppException;
 import alimentation.cashierApp.models.Report;
 
 @Service
@@ -14,29 +16,39 @@ public class ReportService {
     @Autowired
     private ReportRepository ReportRepository;
 
-
     // Add methods for Report-related operations here
-    Iterable<Report> getAllReports(){
+    public Iterable<Report> getAllReports() {
         return ReportRepository.findAll();
     }
 
-    Report getReportById(int id){
+    public Report getReportById(int id) {
         Optional<Report> response = ReportRepository.findById(id);
-        if (response.isPresent()) {
-            return response.get();
+        if (response.isEmpty()) {
+            throw new CashierAppException(HttpStatus.BAD_REQUEST, "Report missing");
         }
-        return null;
+        return response.get();
     }
 
-    void addReport(Report Report){
-        ReportRepository.save(Report);
+    public Report createReport(Report report){
+        // error catchihng implemented in service layer temporarily, 
+        // will be moved to a filtering, processing layer
+        if (report.getEmployee() == null) {
+            throw new CashierAppException(HttpStatus.BAD_REQUEST, "Employee missing");
+        }
+        
+        return ReportRepository.save(report); //internal server error to be handled
+
     }
 
-    void updateReport(Report Report){
-        ReportRepository.save(Report);
+    public Report updateReport(Report report) {
+        return ReportRepository.save(report);
     }
 
-    void deleteReport(int id){
+    public void deleteReport(int id) {
+        Optional<Report> target = ReportRepository.findById(id);
+        if (target.isEmpty()) {
+            throw new CashierAppException(HttpStatus.BAD_REQUEST, "Report missing");
+        }
         ReportRepository.deleteById(id);
     }
 }
